+++ pkg/runtime/containerd.go
@@ -9,10 +9,10 @@ import (
 	"github.com/containerd/containerd"
 	"github.com/containerd/containerd/namespaces"
 
-	"iops-limit-service/pkg/cgroup"
-	"iops-limit-service/pkg/config"
-	"iops-limit-service/pkg/container"
-	"iops-limit-service/pkg/device"
+	"KubeDiskGuard/pkg/cgroup"
+	"KubeDiskGuard/pkg/config"
+	"KubeDiskGuard/pkg/container"
+	"KubeDiskGuard/pkg/device"
 )
 
 // ContainerdRuntime containerd运行时
@@ -108,22 +108,22 @@ func (c *ContainerdRuntime) ProcessContainer(container *container.ContainerInfo)
 	return nil
 }
 
-// SetIOPSLimit 动态设置IOPS限制
-func (c *ContainerdRuntime) SetIOPSLimit(container *container.ContainerInfo, iopsLimit int) error {
+// SetLimits 统一设置IOPS和BPS限制
+func (c *ContainerdRuntime) SetLimits(container *container.ContainerInfo, riops, wiops, rbps, wbps int) error {
 	majMin, err := device.GetMajMin(c.config.DataMount)
 	if err != nil {
 		return err
 	}
 	cgroupPath := c.cgroup.FindCgroupPath(container.ID)
-	return c.cgroup.SetIOPSLimit(cgroupPath, majMin, iopsLimit)
+	return c.cgroup.SetLimits(cgroupPath, majMin, riops, wiops, rbps, wbps)
 }
 
-// ResetIOPSLimit 解除IOPS限制
-func (c *ContainerdRuntime) ResetIOPSLimit(container *container.ContainerInfo) error {
+// ResetLimits 统一解除所有限速
+func (c *ContainerdRuntime) ResetLimits(container *container.ContainerInfo) error {
 	majMin, err := device.GetMajMin(c.config.DataMount)
 	if err != nil {
 		return err
 	}
 	cgroupPath := c.cgroup.FindCgroupPath(container.ID)
-	return c.cgroup.ResetIOPSLimit(cgroupPath, majMin)
+	return c.cgroup.ResetLimits(cgroupPath, majMin)
 }